Updated version

+++ api/routes/users.js
@@ -6,11 +6,12 @@ const bcrypt = require("bcrypt"); //it hashes my password to make it more safety
 //UPDATE
 router.put("/:id", async (req, res) => {
   if (req.body.userId === req.params.id) {
-    if (req.body.password) {
-      const salt = await bcrypt.genSalt(10);
-      req.body.password = await bcrypt.hash(req.body.password, salt);
-    }
     try {
+      if (req.body.password) {
+        const salt = await bcrypt.genSalt(10);
+        req.body.password = await bcrypt.hash(req.body.password, salt);
+      }
+
       const updatedUser = await User.findByIdAndUpdate(
         req.params.id,
         {
@@ -19,7 +20,8 @@ router.put("/:id", async (req, res) => {
         { new: true } //To Show the changes in real time
       );
 
-      res.status(200).json(updatedUser);
+      const { password, ...others } = updatedUser._doc;
+      res.status(200).json(others);
     } catch (error) {
       res.status(500).json(error);
     }
@@ -33,15 +35,16 @@ router.delete("/:id", async (req, res) => {
   if (req.body.userId === req.params.id) {
     try {
       const user = await User.findById(req.params.id);
-      try {
+
+      if (!user) {
+        res.status(401).json("User not found!");
+      } else {
         await Post.deleteMany({ username: user.username });
         await User.findByIdAndDelete(req.params.id);
-        res.status(200).json("User has been deleted!");
-      } catch (error) {
-        res.status(500).json(error);
+        res.status(200).json("User has been deleted successfuly!");
       }
     } catch (error) {
-      res.status(401).json("User not found!");
+      es.status(500).json(error);
     }
   } else {
     res.status(401).json("You can delete only your account!");
@@ -52,8 +55,12 @@ router.delete("/:id", async (req, res) => {
 router.get("/:id", async (req, res) => {
   try {
     const user = await User.findById(req.params.id);
-    const { password, ...others } = user._doc;
-    res.status(200).json(others);
+    if (user) {
+      const { password, ...others } = user._doc;
+      res.status(200).json(others);
+    } else {
+      res.status(404).json("User not found!");
+    }
   } catch (error) {
     res.status(500).json(error);
   }
